+++ js/views/memoryChallenge.js
@@ -1,0 +1,4 @@
+import {Game} from "../models/modelMemoryCard.js"
+
+const game = new Game();
+game.init();
